Move AlloyDBClusterSpec mapper functions to cluster_mappings.go to avoid being overwritten by generator
