+++ pkg/controller/direct/alloydb/mapper.generated.go
@@ -218,98 +218,6 @@ func AlloyDBClusterObservedState_ToProto(mapCtx *direct.MapContext, in *krm.Allo
 	// MISSING: ServiceAccountEmail
 	return out
 }
-func AlloyDBClusterSpec_FromProto(mapCtx *direct.MapContext, in *pb.Cluster) *krm.AlloyDBClusterSpec {
-	if in == nil {
-		return nil
-	}
-	out := &krm.AlloyDBClusterSpec{}
-	// MISSING: BackupSource
-	// MISSING: MigrationSource
-	// MISSING: CloudsqlBackupRunSource
-	// MISSING: Name
-	out.DisplayName = direct.LazyPtr(in.GetDisplayName())
-	// MISSING: Uid
-	// MISSING: CreateTime
-	// MISSING: UpdateTime
-	// MISSING: DeleteTime
-	// MISSING: Labels
-	// MISSING: State
-	out.ClusterType = direct.Enum_FromProto(mapCtx, in.GetClusterType())
-	out.DatabaseVersion = direct.Enum_FromProto(mapCtx, in.GetDatabaseVersion())
-	out.NetworkConfig = Cluster_NetworkConfig_FromProto(mapCtx, in.GetNetworkConfig())
-	if in.GetNetwork() != "" {
-		out.NetworkRef = &krmcomputev1beta1.ComputeNetworkRef{External: in.GetNetwork()}
-	}
-	// MISSING: Etag
-	// MISSING: Annotations
-	// MISSING: Reconciling
-	out.InitialUser = UserPassword_FromProto(mapCtx, in.GetInitialUser())
-	out.AutomatedBackupPolicy = AutomatedBackupPolicy_FromProto(mapCtx, in.GetAutomatedBackupPolicy())
-	// MISSING: SSLConfig
-	out.EncryptionConfig = EncryptionConfig_FromProto(mapCtx, in.GetEncryptionConfig())
-	// MISSING: EncryptionInfo
-	out.ContinuousBackupConfig = ContinuousBackupConfig_FromProto(mapCtx, in.GetContinuousBackupConfig())
-	// MISSING: ContinuousBackupInfo
-	out.SecondaryConfig = Cluster_SecondaryConfig_FromProto(mapCtx, in.GetSecondaryConfig())
-	// MISSING: PrimaryConfig
-	// MISSING: SatisfiesPzs
-	// MISSING: PSCConfig
-	out.MaintenanceUpdatePolicy = MaintenanceUpdatePolicy_FromProto(mapCtx, in.GetMaintenanceUpdatePolicy())
-	// MISSING: MaintenanceSchedule
-	// MISSING: GeminiConfig
-	// MISSING: SubscriptionType
-	// MISSING: TrialMetadata
-	// MISSING: Tags
-	// MISSING: ServiceAccountEmail
-	out.DataplexConfig = Cluster_DataplexConfig_FromProto(mapCtx, in.GetDataplexConfig())
-	return out
-}
-func AlloyDBClusterSpec_ToProto(mapCtx *direct.MapContext, in *krm.AlloyDBClusterSpec) *pb.Cluster {
-	if in == nil {
-		return nil
-	}
-	out := &pb.Cluster{}
-	// MISSING: BackupSource
-	// MISSING: MigrationSource
-	// MISSING: CloudsqlBackupRunSource
-	// MISSING: Name
-	out.DisplayName = direct.ValueOf(in.DisplayName)
-	// MISSING: Uid
-	// MISSING: CreateTime
-	// MISSING: UpdateTime
-	// MISSING: DeleteTime
-	// MISSING: Labels
-	// MISSING: State
-	out.ClusterType = direct.Enum_ToProto[pb.Cluster_ClusterType](mapCtx, in.ClusterType)
-	out.DatabaseVersion = direct.Enum_ToProto[pb.DatabaseVersion](mapCtx, in.DatabaseVersion)
-	out.NetworkConfig = Cluster_NetworkConfig_ToProto(mapCtx, in.NetworkConfig)
-	if in.NetworkRef != nil {
-		out.Network = in.NetworkRef.External
-	}
-	// MISSING: Etag
-	// MISSING: Annotations
-	// MISSING: Reconciling
-	out.InitialUser = UserPassword_ToProto(mapCtx, in.InitialUser)
-	out.AutomatedBackupPolicy = AutomatedBackupPolicy_ToProto(mapCtx, in.AutomatedBackupPolicy)
-	// MISSING: SSLConfig
-	out.EncryptionConfig = EncryptionConfig_ToProto(mapCtx, in.EncryptionConfig)
-	// MISSING: EncryptionInfo
-	out.ContinuousBackupConfig = ContinuousBackupConfig_ToProto(mapCtx, in.ContinuousBackupConfig)
-	// MISSING: ContinuousBackupInfo
-	out.SecondaryConfig = Cluster_SecondaryConfig_ToProto(mapCtx, in.SecondaryConfig)
-	// MISSING: PrimaryConfig
-	// MISSING: SatisfiesPzs
-	// MISSING: PSCConfig
-	out.MaintenanceUpdatePolicy = MaintenanceUpdatePolicy_ToProto(mapCtx, in.MaintenanceUpdatePolicy)
-	// MISSING: MaintenanceSchedule
-	// MISSING: GeminiConfig
-	// MISSING: SubscriptionType
-	// MISSING: TrialMetadata
-	// MISSING: Tags
-	// MISSING: ServiceAccountEmail
-	out.DataplexConfig = Cluster_DataplexConfig_ToProto(mapCtx, in.DataplexConfig)
-	return out
-}
 func AlloyDBInstanceObservedState_FromProto(mapCtx *direct.MapContext, in *pb.Instance) *krm.AlloyDBInstanceObservedState {
 	if in == nil {
 		return nil
@@ -576,22 +484,6 @@ func CloudSQLBackupRunSource_ToProto(mapCtx *direct.MapContext, in *krm.CloudSQL
 	out.Project = direct.ValueOf(in.Project)
 	out.InstanceId = direct.ValueOf(in.InstanceID)
 	out.BackupRunId = direct.ValueOf(in.BackupRunID)
-	return out
-}
-func Cluster_DataplexConfig_FromProto(mapCtx *direct.MapContext, in *pb.Cluster_DataplexConfig) *krm.Cluster_DataplexConfig {
-	if in == nil {
-		return nil
-	}
-	out := &krm.Cluster_DataplexConfig{}
-	out.Enabled = direct.LazyPtr(in.GetEnabled())
-	return out
-}
-func Cluster_DataplexConfig_ToProto(mapCtx *direct.MapContext, in *krm.Cluster_DataplexConfig) *pb.Cluster_DataplexConfig {
-	if in == nil {
-		return nil
-	}
-	out := &pb.Cluster_DataplexConfig{}
-	out.Enabled = direct.ValueOf(in.Enabled)
 	return out
 }
 func Cluster_NetworkConfig_FromProto(mapCtx *direct.MapContext, in *pb.Cluster_NetworkConfig) *krm.Cluster_NetworkConfig {
@@ -881,7 +773,6 @@ func Instance_ConnectionPoolConfig_FromProto(mapCtx *direct.MapContext, in *pb.I
 	out := &krm.Instance_ConnectionPoolConfig{}
 	out.Enabled = direct.LazyPtr(in.GetEnabled())
 	out.Flags = in.Flags
-	// MISSING: PoolerCount
 	return out
 }
 func Instance_ConnectionPoolConfig_ToProto(mapCtx *direct.MapContext, in *krm.Instance_ConnectionPoolConfig) *pb.Instance_ConnectionPoolConfig {
@@ -891,7 +782,6 @@ func Instance_ConnectionPoolConfig_ToProto(mapCtx *direct.MapContext, in *krm.In
 	out := &pb.Instance_ConnectionPoolConfig{}
 	out.Enabled = direct.ValueOf(in.Enabled)
 	out.Flags = in.Flags
-	// MISSING: PoolerCount
 	return out
 }
 func Instance_InstanceNetworkConfig_FromProto(mapCtx *direct.MapContext, in *pb.Instance_InstanceNetworkConfig) *krm.Instance_InstanceNetworkConfig {
